+++ js/const.js
@@ -1,0 +1,22 @@
+const AUTHOR_NAMES = [
+  'Дмитрий',
+  'Евгений',
+  'Максим',
+  'Сергей',
+  'Анна',
+  'Валерий',
+];
+
+const LIKES = {
+  min: 5,
+  max: 10,
+};
+
+const COMMENTS = {
+  min: 1,
+  max: 5,
+};
+
+const COMMENTS_LIST = ['Всё отлично!', 'В целом всё неплохо. Но не всё.'];
+
+export { AUTHOR_NAMES, LIKES, COMMENTS, COMMENTS_LIST };
